requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
chardet==5.2.0
colorama==0.4.6
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
numpy==2.1.0
pandas==2.2.2
pycodestyle==2.12.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.5
tzdata==2024.1
ujson==5.10.0
yarl==1.9.7
pyotp-2.9.0